requirements.txt
nextcord==2.5.0
python-dotenv==1.0.0
deep-translator==1.11.4
googlesearch-python==1.2.3
pyjokes==0.6.0
